Delete button component and back button component
