+++ stockflux-securities-master/src/components/form-view/form-title/FormTitle.css
@@ -1,0 +1,4 @@
+.title {
+  padding: var(--padding-large);
+  font-weight: var(--font-weight-bold);
+}
