+++ stockflux-securities-master/src/components/securities/table/body/Body.css
@@ -1,7 +1,7 @@
 .tbody {
   display: grid;
 }
-.scrollable-table-body {
+.scrollable {
   max-height: calc(600px - 54px - 35px - 36px - 4px);
   overflow: overlay;
 }
